Major refactor
Added saturation parameter since we are simply boosting highlights, and so highlights very quickly get clipped to white without it. 

Changed the way luminance and gain is calculated, so with neutral 1.0 values, motion blur will no longer over brighten highlights and mids, causing flickering when there is a lot of bright pixels.